Convergence + optimizations
Pixel colour converges and exits early if the change is below a threshold. BVH intersection also exits early if it misses the entire structure, which skips queue setup and the 2 child intersections, saving us 1 AABB test. On open sky models a lot of diffusion samples and reflections hit the sky so this improves performance slightly.